Refactor isatab module to separate metadata and data parsing
